Use nova-consoleauth only if workaround enabled
In Rocky, we deprecated the nova-consoleauth service but there were
unconditional calls to nova-consoleauth in the compute/api, which
made it impossible to avoid running the nova-consoleauth service.

This adds conditional checks to call nova-consoleauth only if the
[workarounds]enable_consoleauth configuration option is True. The
option defaults to False because the default console token auth TTL
is 10 minutes and only operators who have configured much longer TTL
or otherwise wish to avoid resetting all consoles at upgrade time
need to use the option.

This also updates the /os-console-auth-tokens/{console_token} API to
use nova-consoleauth only if the [workarounds] option is enabled. This
had to be done in the same change because the conditional checks in
the compute/api code caused the /os-console-auth-tokens API functional
tests to fail to find token authorizations in nova-consoleauth.

Closes-Bug: #1788470
Closes-Bug: #1795982

(cherry picked from commit b49209cc29f5ccd529b8074b68be074c5a2f742f)

+++ nova/compute/api.py
@@ -1944,12 +1944,11 @@ class API(base.Base):
 
             # NOTE(dtp): cells.enable = False means "use cells v2".
             # Run everywhere except v1 compute cells.
-            if not CONF.cells.enable or self.cell_type == 'api':
-                # TODO(melwitt): In Rocky, we store console authorizations
-                # in both the consoleauth service and the database while
-                # we convert to using the database. Remove the consoleauth
-                # line below when authorizations are no longer being
-                # stored in consoleauth, in Stein.
+            if (not CONF.cells.enable and CONF.workarounds.enable_consoleauth
+               ) or self.cell_type == 'api':
+                # TODO(melwitt): Remove the conditions for running this line
+                # with cells v2, when consoleauth is no longer being used by
+                # cells v2, in Stein.
                 self.consoleauth_rpcapi.delete_tokens_for_instance(
                     context, instance.uuid)
 
@@ -3743,11 +3742,12 @@ class API(base.Base):
         # console authorization in the database in the above method.
         # The following will be removed when everything has been
         # converted to use the database, in Stein.
-        self.consoleauth_rpcapi.authorize_console(context,
-                connect_info['token'], console_type,
-                connect_info['host'], connect_info['port'],
-                connect_info['internal_access_path'], instance.uuid,
-                access_url=connect_info['access_url'])
+        if CONF.workarounds.enable_consoleauth:
+            self.consoleauth_rpcapi.authorize_console(context,
+                    connect_info['token'], console_type,
+                    connect_info['host'], connect_info['port'],
+                    connect_info['internal_access_path'], instance.uuid,
+                    access_url=connect_info['access_url'])
 
         return {'url': connect_info['access_url']}
 
@@ -3769,11 +3769,12 @@ class API(base.Base):
         # console authorization in the database in the above method.
         # The following will be removed when everything has been
         # converted to use the database, in Stein.
-        self.consoleauth_rpcapi.authorize_console(context,
-                connect_info['token'], console_type,
-                connect_info['host'], connect_info['port'],
-                connect_info['internal_access_path'], instance.uuid,
-                access_url=connect_info['access_url'])
+        if CONF.workarounds.enable_consoleauth:
+            self.consoleauth_rpcapi.authorize_console(context,
+                    connect_info['token'], console_type,
+                    connect_info['host'], connect_info['port'],
+                    connect_info['internal_access_path'], instance.uuid,
+                    access_url=connect_info['access_url'])
 
         return {'url': connect_info['access_url']}
 
@@ -3795,11 +3796,12 @@ class API(base.Base):
         # console authorization in the database in the above method.
         # The following will be removed when everything has been
         # converted to use the database, in Stein.
-        self.consoleauth_rpcapi.authorize_console(context,
-                connect_info['token'], console_type,
-                connect_info['host'], connect_info['port'],
-                connect_info['internal_access_path'], instance.uuid,
-                access_url=connect_info['access_url'])
+        if CONF.workarounds.enable_consoleauth:
+            self.consoleauth_rpcapi.authorize_console(context,
+                    connect_info['token'], console_type,
+                    connect_info['host'], connect_info['port'],
+                    connect_info['internal_access_path'], instance.uuid,
+                    access_url=connect_info['access_url'])
 
         return {'url': connect_info['access_url']}
 
@@ -3822,11 +3824,12 @@ class API(base.Base):
         # console authorization in the database in the above method.
         # The following will be removed when everything has been
         # converted to use the database, in Stein.
-        self.consoleauth_rpcapi.authorize_console(context,
-                connect_info['token'], console_type,
-                connect_info['host'], connect_info['port'],
-                connect_info['internal_access_path'], instance.uuid,
-                access_url=connect_info['access_url'])
+        if CONF.workarounds.enable_consoleauth:
+            self.consoleauth_rpcapi.authorize_console(context,
+                    connect_info['token'], console_type,
+                    connect_info['host'], connect_info['port'],
+                    connect_info['internal_access_path'], instance.uuid,
+                    access_url=connect_info['access_url'])
         return {'url': connect_info['access_url']}
 
     @check_instance_host
@@ -3847,11 +3850,12 @@ class API(base.Base):
         # console authorization in the database in the above method.
         # The following will be removed when everything has been
         # converted to use the database, in Stein.
-        self.consoleauth_rpcapi.authorize_console(context,
-                connect_info['token'], console_type,
-                connect_info['host'], connect_info['port'],
-                connect_info['internal_access_path'], instance.uuid,
-                access_url=connect_info['access_url'])
+        if CONF.workarounds.enable_consoleauth:
+            self.consoleauth_rpcapi.authorize_console(context,
+                    connect_info['token'], console_type,
+                    connect_info['host'], connect_info['port'],
+                    connect_info['internal_access_path'], instance.uuid,
+                    access_url=connect_info['access_url'])
         return {'url': connect_info['access_url']}
 
     @check_instance_host
@@ -4376,13 +4380,14 @@ class API(base.Base):
         self._record_action_start(context, instance,
                                   instance_actions.LIVE_MIGRATION)
 
-        # TODO(melwitt): In Rocky, we store console authorizations
+        # TODO(melwitt): In Rocky, we optionally store console authorizations
         # in both the consoleauth service and the database while
-        # we convert to using the database. Remove the consoleauth
-        # line below when authorizations are no longer being
-        # stored in consoleauth, in Stein.
-        self.consoleauth_rpcapi.delete_tokens_for_instance(
-            context, instance.uuid)
+        # we convert to using the database. Remove the condition for running
+        # this line with cells v2, when consoleauth is no longer being used by
+        # cells v2, in Stein.
+        if CONF.cells.enable or CONF.workarounds.enable_consoleauth:
+            self.consoleauth_rpcapi.delete_tokens_for_instance(
+                context, instance.uuid)
 
         try:
             request_spec = objects.RequestSpec.get_by_instance_uuid(
